requirements.txt
CacheControl==0.13.1
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==41.0.1
firebase-admin==6.1.0
google-api-core==2.11.1
google-api-python-client==2.89.0
google-auth==2.20.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.11.1
google-cloud-storage==2.9.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.1
grpcio==1.54.2
grpcio-status==1.54.2
httplib2==0.22.0
idna==3.4
msgpack==1.0.5
proto-plus==1.22.2
protobuf==4.23.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.9
PyJWT==2.7.0
pyparsing==3.0.9
python-dotenv==1.0.0
requests==2.31.0
rsa==4.9
six==1.16.0
typing_extensions==4.6.3
uritemplate==4.1.1
urllib3==1.26.16
